Transmitter operated mice

Data from a transmitter implantation experiment with various behavioural data. Also includes sham-operated animals. Observation time was over 30 days, starting with day=-1 as baseline (physiological state). The treatment columns shows what operational state the animal is in and the condition field shows two possible types of analgesia.

data(postop)

Format

A 840 x 16 data frame with post-op data.

Examples

data(postop)
head(postop)

#>       id   treatment condition day   bw       bwc score      mgs     mgs30
#> 1 Ca_001 Transmitter Carprofen  -1 18.4 100.00000     0 0.100000        NA
#> 2 Ca_001 Transmitter Carprofen   0 16.1  87.50000     2 0.112500 0.8611607
#> 3 Ca_001 Transmitter Carprofen   1 16.7  90.76087     2 0.193750        NA
#> 4 Ca_001 Transmitter Carprofen   2 17.1  92.93478     2 0.165625        NA
#> 5 Ca_001 Transmitter Carprofen   3 17.3  94.02174     1       NA        NA
#> 6 Ca_001 Transmitter Carprofen   4 17.3  94.02174     1 0.065625        NA
#>   mgs180    bur2h    burON       hr       hrv     temp     act
#> 1     NA 56.39602 71.94116 453.5596 12.432154 36.61165 1064.84
#> 2 0.3625 14.92109 44.69154 660.5897  2.051861 36.18393  125.54
#> 3     NA 51.07296 97.63948 584.9174  4.843867 37.49761  407.34
#> 4     NA 42.28045 99.00850 538.0367  6.767956 37.13743  351.64
#> 5     NA       NA       NA 482.1560  7.072865 36.71055  286.28
#> 6     NA 54.90473 69.16020 525.5229  7.134940 37.08358  387.30