RELSA Wrapper

The relsa_wrapper analyses the data set and calculates various RELSA objects for later analysis. Please note that the data must be in the RELSA format.

relsa_wrapper(
  querydata,
  baseline = NULL,
  treatment = NULL,
  condition = NULL,
  normthese = NULL,
  turnsQuery = NULL,
  dropsQuery = NULL,
  animalnr = 1,
  ymax = 1.2,
  pcadims = 2,
  studylabel = NULL,
  severity = NULL,
  colorlabel = NULL
)

Arguments

querydata
dataset sample data frame

baseline
cluster baselines from relsa_baselines function

treatment
treatment column, potential data filter for subsampling (use column name as character, e.g. "transmitter")

condition
condition column, potential data filter for subsampling (use column name as character, e.g. "Carprofen")

normthese
variable vector (colum names as character, e.g. c("hr","temp")) with the names of the variables that need "normalization"

turnsQuery
variable vector (colum names as character, e.g. c("hr","temp")) with the names of the variables that need "turning"

dropsQuery
variables to drop

animalnr
animal number (not id) in the data set for which an example RELSA flow is generated (defaults to 1)

ymax
y-max of the y axis (defaults to 1.2)

pcadims
for how many principal components shall the dimensional contributions be calculated (note: only 2 can be plotted, defaults to 2)

studylabel
you can assign an individual study name to label some of the tables in the relsa object

severity
you can assign a prospective severity for your experiment (has no effect on model calculations, just information)

colorlabel
you can assign a color label to some tables of the relsa object

Value

robj RELSA object - a list with RELSA calculation outputs.